% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.factorAnalysis}
\alias{plot.factorAnalysis}
\title{Plot a Factor Solution}
\usage{
\method{plot}{factorAnalysis}(x, y, ..., sort = FALSE, type = c("bar", "table"))
}
\arguments{
\item{x}{an object of class \code{factorAnalysis} produced by the \code{\link{PCA}}
or \code{\link{FA}} functions.}

\item{y}{not used}

\item{...}{not used}

\item{sort}{logical. If \code{TRUE}, sort the pattern matrix.}

\item{type}{generate a bar plot (\code{"bar"}) or table plot (\code{"table"}).}
}
\value{
a ggplot2 graph
}
\description{
This function plots the pattern matrix for a principal components or
common factor analysis solution
}
\examples{
fit.pca <- PCA(Harman74.cor$cov, nfactors=4, rotate="varimax")
plot(fit.pca, sort=TRUE)
plot(fit.pca, sort=TRUE, type="table")
}
