% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacDR-package.R
\docType{package}
\name{qacDR-package}
\alias{qacDR}
\alias{qacDR-package}
\title{qacDR: Principal Components, Factor Analysis, and Clustering Tools}
\description{
Perform principal components analysis and iterated principal axis factor analysis with visualizations. There are additional tools for use in cluster analysis.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/rkabacoff/qacDR/issues}
}

}
\author{
\strong{Maintainer}: Rob Kabacoff \email{rkabacoff@wesleyan.edu}

}
\keyword{internal}
